requirements.txt
elasticsearch
requests
python-dotenv
InquirerPy
tabulate